# Add substrings search path for constant like pattern