We noticed that some like constant-patterns can be equivalent to substrings search such as tpch queries Q13 and Q16: "%special%requests%" "%Customer%Complaints%"

1.  Add substring-search fast path for like with search pattern : "((%+[^%#_\\]+)+%+)"
2. Some OLAP engine supports "\_" for substring-search. In order to obtain the best performance, "\_" is not supported here, '#' is not support for the same reason too.
3. This PR will also affect single substring-search like tpch Q9,  if a fast-simd-strstr to speedup string-search will be added as follow-up, changes nothing right now.

benchmark result :
before:

tpchQuery9                                                 23.96ms     41.73
tpchQuery13                                               156.02ms      6.41
tpchQuery16Part                                             9.02ms    110.83
tpchQuery16Supplier                                       192.13ms      5.20

after:

tpchQuery9                                                 24.30ms     41.14
tpchQuery13                                                55.03ms     18.17
tpchQuery16Part                                             8.92ms    112.07
tpchQuery16Supplier                                        17.64ms     56.69