@skadilover 2024-08-13
@mbasmanova  @Yuhta 

Could you help to have a look at this PR ? The code format is still being fixed, I am not sure is it reasonable to add a fast-path here.
And could you help to look at the "load_aligned" method use in smid_strstr_memcmp is safe as it may load some unkown-memory address here.