@skadilover 2024-08-14
@Yuhta  update code , could you have a look at this once more ?

> i + CharVector::size <= n otherwise it's overreading, you need to handle trailing characters separately

I think this is the most interesting thing in this PR:
1. I add pageSafe check here now , like  ClickHouse does : https://github.com/ClickHouse/ClickHouse/blob/a51f867ce014a4cb8f97c46e004b90f5feafd80f/src/Common/StringSearcher.h#L31
```
    bool isPageSafe(const void * const ptr) const
    {
        return ((page_size - 1) & reinterpret_cast<std::uintptr_t>(ptr)) <= page_size - N;
    }
```
By this way,  load_unaligned is safe when overreading for they are in same memory page.

2. For StringSearch, ClickHouse use Volnitsky  https://github.com/ClickHouse/ClickHouse/blob/1513eda3efe3a24b7a87b9f0e7803bb8693d1bc8/src/Common/Volnitsky.h#L373 

Howevever, performance seems not best, Doris delete Volnitsky when they copy the StringSearch implement : https://github.com/apache/doris/blob/4e326e5f496109edb253d094869d057573ab7a59/be/src/vec/common/string_searcher.h#L44

And use first-two char compare, however, I test it , about 2x slow than this PR`s implement.