@skadilover 2024-08-16
@Yuhta  ping here~ This PR is relative smaller and cheaper,  however, affect tpch-Q13`s performance a lot, about 2.5% up for total TPCH test.