@mbasmanova 2024-08-19
@skadilover Are there 2 separate optimizations here? Can SIMD changes be extracted into a separate PR?

LIKE is a widely used construct. We need to make sure the change is 100% correct. Would you describe the testing approach and why do you believe it guarantees 100% correctness?

CC: @kagamiori @kgpai @amitkdutta