@mbasmanova 2024-08-19
@skadilover Please, make sure to document the new optimizations in https://facebookincubator.github.io/velox/functions/presto/regexp.html