@mbasmanova 2024-09-11
> I will add benchmark to measure determinePatternKind(), and disscuss it again, I simplely assume that parseSubstrings will add perf regression and only add it for constant-case.

@skadilover What did you find?