@skadilover 2024-09-12
@mbasmanova  update codes:
1.  update like documentation. 
2.  fix comments for kSubstrings as you mentioned.
3.  Regarding the problem of ‘determinePatternKind()’, let’s ignore it for now. I will study it clearly to see if it has performance problems in non-constant situations after I finish handling some existing work.

>I will add benchmark to measure determinePatternKind(), and disscuss it again, I simplely assume that parseSubstrings will add perf regression and only add it for constant-case.

>@skadilover What did you find?