@Yuhta 2024-09-20
Is the regression in benchmark caused by the change?  If so we probably want to merge this after #10858