@skadilover 2024-09-20
> Is the regression in benchmark caused by the change? If so we probably want to merge this after #10858

It`s strange because Q2 and suffix is not supported by this change.